ProblemStatement1/requirements.txt
fastapi
uvicorn
pydantic
openai
groq
python-dotenv
langchain
langchain_community